Updated UI, removed code that had no use and added comments
