+++ lib/notifiers/progress_notifier.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-// blog and github for code *** EDIT THIS CODE ***
+// blog and github for code
 // https://suragch.medium.com/managing-playlists-in-flutter-with-just-audio-c4b8f2af12eb
 // https://github.com/suragch/audio_playlist_flutter_demo
