--- lib/pages/player.dart
@@ -1,397 +1,0 @@
-import 'package:audio_video_progress_bar/audio_video_progress_bar.dart';
-import 'package:flutter/material.dart';
-
-import 'package:just_audio/just_audio.dart';
-import 'package:flutter_project_2022/notifiers/play_button_notifier.dart';
-import 'package:flutter_project_2022/notifiers/progress_notifier.dart';
-import 'package:flutter_project_2022/notifiers/repeat_button_notifier.dart';
-//import 'pageManager.dart';
-
-// blog and github for code *** EDIT THIS CODE *** *** Get music playing ***
-// https://suragch.medium.com/managing-playlists-in-flutter-with-just-audio-c4b8f2af12eb
-// https://github.com/suragch/audio_playlist_flutter_demo
-
-class Player extends StatefulWidget {
-  final String artist;
-  final String title;
-  final String url;
-  final String image;
-
-  Player(
-      {required this.title,
-      required this.artist,
-      required this.url,
-      required this.image});
-
-  @override
-  _PlayerState createState() => _PlayerState();
-}
-
-// use GetIt or Provider rather than a global variable in a real project
-late final PageManager _pageManager;
-
-class _PlayerState extends State<Player> {
-  @override
-  void initState() {
-    super.initState();
-    _init();
-    _pageManager = PageManager();
-  }
-
-  @override
-  void dispose() {
-    _pageManager.dispose();
-    super.dispose();
-  }
-
-  static const url =
-      "https://firebasestorage.googleapis.com/v0/b/flutter-project-2022.appspot.com/o/songs%2FQvaJnwom58SD4WGsXZChE8s2ezK2%2FEDEN%20-%20xo%20(official%20audio).mp3?alt=media&token=98fe2a78-7a94-4ebc-8b0f-ca582cffe276";
-
-  late AudioPlayer _audioPlayer;
-
-  void _init() async {
-    _audioPlayer = AudioPlayer();
-    await _audioPlayer.setUrl(url);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text("Playing"),
-        ),
-        body: Padding(
-          padding: const EdgeInsets.all(20.0),
-          child: Column(
-            children: [
-              CurrentSongTitle(),
-              AudioProgressBar(),
-              AudioControlButtons(),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class CurrentSongTitle extends StatelessWidget {
-  const CurrentSongTitle({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<String>(
-      valueListenable: _pageManager.currentSongTitleNotifier,
-      builder: (_, title, __) {
-        return Padding(
-          padding: const EdgeInsets.only(top: 8.0),
-          child: Text(title, style: TextStyle(fontSize: 40)),
-        );
-      },
-    );
-  }
-}
-
-class AudioProgressBar extends StatelessWidget {
-  const AudioProgressBar({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<ProgressBarState>(
-      valueListenable: _pageManager.progressNotifier,
-      builder: (_, value, __) {
-        return ProgressBar(
-          progress: value.current,
-          buffered: value.buffered,
-          total: value.total,
-          onSeek: _pageManager.seek,
-        );
-      },
-    );
-  }
-}
-
-class AudioControlButtons extends StatelessWidget {
-  const AudioControlButtons({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 60,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: [
-          RepeatButton(),
-          PreviousSongButton(),
-          PlayButton(),
-          NextSongButton(),
-          ShuffleButton(),
-        ],
-      ),
-    );
-  }
-}
-
-class RepeatButton extends StatelessWidget {
-  const RepeatButton({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<RepeatState>(
-      valueListenable: _pageManager.repeatButtonNotifier,
-      builder: (context, value, child) {
-        Icon icon;
-        switch (value) {
-          case RepeatState.off:
-            icon = Icon(Icons.repeat, color: Colors.grey);
-            break;
-          case RepeatState.repeatSong:
-            icon = Icon(Icons.repeat_one);
-            break;
-          case RepeatState.repeatPlaylist:
-            icon = Icon(Icons.repeat);
-            break;
-        }
-        return IconButton(
-          icon: icon,
-          onPressed: _pageManager.onRepeatButtonPressed,
-        );
-      },
-    );
-  }
-}
-
-class PreviousSongButton extends StatelessWidget {
-  const PreviousSongButton({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<bool>(
-      valueListenable: _pageManager.isFirstSongNotifier,
-      builder: (_, isFirst, __) {
-        return IconButton(
-          icon: Icon(Icons.skip_previous),
-          onPressed:
-              (isFirst) ? null : _pageManager.onPreviousSongButtonPressed,
-        );
-      },
-    );
-  }
-}
-
-class PlayButton extends StatelessWidget {
-  const PlayButton({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<ButtonState>(
-      valueListenable: _pageManager.playButtonNotifier,
-      builder: (_, value, __) {
-        switch (value) {
-          case ButtonState.loading:
-            return Container(
-              margin: EdgeInsets.all(8.0),
-              width: 32.0,
-              height: 32.0,
-              child: CircularProgressIndicator(),
-            );
-          case ButtonState.paused:
-            return IconButton(
-              icon: Icon(Icons.play_arrow),
-              iconSize: 32.0,
-              onPressed: _pageManager.play,
-            );
-          case ButtonState.playing:
-            return IconButton(
-              icon: Icon(Icons.pause),
-              iconSize: 32.0,
-              onPressed: _pageManager.pause,
-            );
-        }
-      },
-    );
-  }
-}
-
-class NextSongButton extends StatelessWidget {
-  const NextSongButton({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<bool>(
-      valueListenable: _pageManager.isLastSongNotifier,
-      builder: (_, isLast, __) {
-        return IconButton(
-          icon: Icon(Icons.skip_next),
-          onPressed: (isLast) ? null : _pageManager.onNextSongButtonPressed,
-        );
-      },
-    );
-  }
-}
-
-class ShuffleButton extends StatelessWidget {
-  const ShuffleButton({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<bool>(
-      valueListenable: _pageManager.isShuffleModeEnabledNotifier,
-      builder: (context, isEnabled, child) {
-        return IconButton(
-          icon: (isEnabled)
-              ? Icon(Icons.shuffle)
-              : Icon(Icons.shuffle, color: Colors.grey),
-          onPressed: _pageManager.onShuffleButtonPressed,
-        );
-      },
-    );
-  }
-}
-
-class PageManager {
-  final currentSongTitleNotifier = ValueNotifier<String>('');
-  final playlistNotifier = ValueNotifier<List<String>>([]);
-  final progressNotifier = ProgressNotifier();
-  final repeatButtonNotifier = RepeatButtonNotifier();
-  final isFirstSongNotifier = ValueNotifier<bool>(true);
-  final playButtonNotifier = PlayButtonNotifier();
-  final isLastSongNotifier = ValueNotifier<bool>(true);
-  final isShuffleModeEnabledNotifier = ValueNotifier<bool>(false);
-
-  static const url = "url";
-
-  late AudioPlayer _audioPlayer;
-
-  PageManager() {
-    _init();
-  }
-
-  void _init() async {
-    _audioPlayer = AudioPlayer();
-    await _audioPlayer.setUrl(url);
-    _listenForChangesInPlayerState();
-    _listenForChangesInPlayerPosition();
-    _listenForChangesInBufferedPosition();
-    _listenForChangesInTotalDuration();
-    _listenForChangesInSequenceState();
-  }
-
-  void _listenForChangesInPlayerState() {
-    _audioPlayer.playerStateStream.listen((playerState) {
-      final isPlaying = playerState.playing;
-      final processingState = playerState.processingState;
-      if (processingState == ProcessingState.loading ||
-          processingState == ProcessingState.buffering) {
-        playButtonNotifier.value = ButtonState.loading;
-      } else if (!isPlaying) {
-        playButtonNotifier.value = ButtonState.paused;
-      } else if (processingState != ProcessingState.completed) {
-        playButtonNotifier.value = ButtonState.playing;
-      } else {
-        _audioPlayer.seek(Duration.zero);
-        _audioPlayer.pause();
-      }
-    });
-  }
-
-  void _listenForChangesInPlayerPosition() {
-    _audioPlayer.positionStream.listen((position) {
-      final oldState = progressNotifier.value;
-      progressNotifier.value = ProgressBarState(
-        current: position,
-        buffered: oldState.buffered,
-        total: oldState.total,
-      );
-    });
-  }
-
-  void _listenForChangesInBufferedPosition() {
-    _audioPlayer.bufferedPositionStream.listen((bufferedPosition) {
-      final oldState = progressNotifier.value;
-      progressNotifier.value = ProgressBarState(
-        current: oldState.current,
-        buffered: bufferedPosition,
-        total: oldState.total,
-      );
-    });
-  }
-
-  void _listenForChangesInTotalDuration() {
-    _audioPlayer.durationStream.listen((totalDuration) {
-      final oldState = progressNotifier.value;
-      progressNotifier.value = ProgressBarState(
-        current: oldState.current,
-        buffered: oldState.buffered,
-        total: totalDuration ?? Duration.zero,
-      );
-    });
-  }
-
-  void _listenForChangesInSequenceState() {
-    _audioPlayer.sequenceStateStream.listen((sequenceState) {
-      if (sequenceState == null) return;
-
-      // update current song title
-      final currentItem = sequenceState.currentSource;
-      final title = currentItem?.tag as String?;
-      currentSongTitleNotifier.value = title ?? '';
-
-      // update playlist
-      final playlist = sequenceState.effectiveSequence;
-      final titles = playlist.map((item) => item.tag as String).toList();
-      playlistNotifier.value = titles;
-
-      // update shuffle mode
-      isShuffleModeEnabledNotifier.value = sequenceState.shuffleModeEnabled;
-
-      // update previous and next buttons
-      if (playlist.isEmpty || currentItem == null) {
-        isFirstSongNotifier.value = true;
-        isLastSongNotifier.value = true;
-      } else {
-        isFirstSongNotifier.value = playlist.first == currentItem;
-        isLastSongNotifier.value = playlist.last == currentItem;
-      }
-    });
-  }
-
-  void play() async {
-    _audioPlayer.play();
-  }
-
-  void pause() {
-    _audioPlayer.pause();
-  }
-
-  void seek(Duration position) {
-    _audioPlayer.seek(position);
-  }
-
-  void dispose() {
-    _audioPlayer.dispose();
-  }
-
-  void onRepeatButtonPressed() {
-    repeatButtonNotifier.nextState();
-    switch (repeatButtonNotifier.value) {
-      case RepeatState.off:
-        _audioPlayer.setLoopMode(LoopMode.off);
-        break;
-      case RepeatState.repeatSong:
-        _audioPlayer.setLoopMode(LoopMode.one);
-        break;
-      case RepeatState.repeatPlaylist:
-        _audioPlayer.setLoopMode(LoopMode.all);
-    }
-  }
-
-  void onPreviousSongButtonPressed() {
-    _audioPlayer.seekToPrevious();
-  }
-
-  void onNextSongButtonPressed() {
-    _audioPlayer.seekToNext();
-  }
-
-  void onShuffleButtonPressed() async {
-    final enable = !_audioPlayer.shuffleModeEnabled;
-    if (enable) {
-      await _audioPlayer.shuffle();
-    }
-    await _audioPlayer.setShuffleModeEnabled(enable);
-  }
-}
